Rename back to staticRoutes
