+++ app/resources/content/components/naechste-schritte.ts
@@ -2,7 +2,7 @@ import {
   ROUTE_DOCUMENTATION,
   ROUTE_METHODS,
   ROUTE_PRECHECK,
-} from "~/resources/routeDefinitions";
+} from "~/resources/staticRoutes";
 import { dedent } from "~/utils/dedentMultilineStrings";
 
 const stepNKR = {
