+++ app/resources/content/methode-it-systeme-erfassen.ts
@@ -5,7 +5,7 @@ import SupportOutlined from "@digitalservicebund/icons/SupportOutlined";
 import {
   ROUTE_METHODS_FIVE_PRINCIPLES,
   ROUTE_SUPPORT,
-} from "~/resources/routeDefinitions";
+} from "~/resources/staticRoutes";
 import { dedent } from "~/utils/dedentMultilineStrings";
 
 export const methodsITSystems = {
