+++ app/resources/content/startseite.ts
@@ -3,7 +3,7 @@ import {
   ROUTE_EXAMPLES,
   ROUTE_INTEROPERABILITY,
   ROUTE_METHODS_FIVE_PRINCIPLES,
-} from "~/resources/routeDefinitions";
+} from "~/resources/staticRoutes";
 import { dedent } from "~/utils/dedentMultilineStrings";
 
 export const index = {
