+++ app/resources/staticRoutes.ts
