+++ app/routes/beispiele.visualisierungen.tsx
@@ -9,7 +9,7 @@ import { examplesVisualisations } from "~/resources/content/beispiele-visualisie
 import {
   ROUTE_EXAMPLES_VISUALISATIONS,
   ROUTE_REGELUNGEN,
-} from "~/resources/routeDefinitions";
+} from "~/resources/staticRoutes";
 import prependMetaTitle from "~/utils/metaTitle";
 import {
   fetchStrapiData,
