+++ tests/e2e/methodPages.spec.ts
@@ -14,7 +14,7 @@ import {
   ROUTE_METHODS_RESPONSIBLE_ACTORS,
   ROUTE_METHODS_TASKS_PROCESSES,
   ROUTE_METHODS_TECHNICAL_FEASIBILITY,
-} from "~/resources/routeDefinitions";
+} from "~/resources/staticRoutes";
 
 test.describe("test method page link flow", () => {
   test("happy path method pages", async ({ page }, testInfo) => {
